Refactor errors into directory
