+++ server/lib/errors/status-error.js
