+++ server/lib/errors/timeout-abort-error.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const ExtendedError = require('./extended-error');
+
+class TimeoutAbortError extends ExtendedError {
+  // ...
+}
+
+module.exports = TimeoutAbortError;
