+++ server/lib/errors/user-closed-connection-abort-error.js
@@ -1,0 +1,9 @@
+'use strict';
+
+const ExtendedError = require('./extended-error');
+
+class UserClosedConnectionAbortError extends ExtendedError {
+  // ...
+}
+
+module.exports = UserClosedConnectionAbortError;
